# accounts-service/src/main/resources/openapi/accounts-api.yaml
openapi: 3.0.3

info:
  title: Accounts Service API
  description: |
    Manages bank accounts lifecycle: creation, retrieval, status management and balance queries.
    This contract is the source of truth — do not modify generated code directly.
  version: 1.0.0
  contact:
    name: Antonio
    url: https://github.com/TonyKerr07

servers:
  - url: http://localhost:8081
    description: Local development

tags:
  - name: Accounts
    description: Account management operations
  - name: Balance
    description: Balance consultation

paths:
  /api/v1/accounts:
    post:
      tags: [ Accounts ]
      operationId: createAccount
      summary: Create a new bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags: [ Accounts ]
      operationId: listAccounts
      summary: List all accounts (paginated)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPageResponse'

  /api/v1/accounts/{id}:
    get:
      tags: [ Accounts ]
      operationId: findAccountById
      summary: Find account by ID
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [ Accounts ]
      operationId: updateAccountStatus
      summary: Update account status (activate, block, close)
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountStatusRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /api/v1/accounts/{id}/balance:
    get:
      tags: [ Balance ]
      operationId: getBalance
      summary: Get account balance
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

  schemas:
    CreateAccountRequest:
      type: object
      required: [ holderName, documentNumber, accountType ]
      properties:
        holderName:
          type: string
          minLength: 3
          maxLength: 150
          example: "João da Silva"
        documentNumber:
          type: string
          description: CPF (11 digits) or CNPJ (14 digits), numbers only
          example: "12345678901"
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          type: number
          format: double
          minimum: 0
          default: 0.0
          example: 1000.00

    UpdateAccountStatusRequest:
      type: object
      required: [ status ]
      properties:
        status:
          $ref: '#/components/schemas/AccountStatus'

    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        holderName:
          type: string
        documentNumber:
          type: string
        accountNumber:
          type: string
        accountType:
          $ref: '#/components/schemas/AccountType'
        status:
          $ref: '#/components/schemas/AccountStatus'
        balance:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AccountPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean

    BalanceResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        accountNumber:
          type: string
        balance:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/AccountStatus'
        consultedAt:
          type: string
          format: date-time

    AccountType:
      type: string
      enum: [ CHECKING, SAVINGS, SALARY ]

    AccountStatus:
      type: string
      enum: [ ACTIVE, BLOCKED, CLOSED ]

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        errorCode:
          type: string
        message:
          type: string
        path:
          type: string

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'